메시지 페이지 작업중 + 기타 자잘한 수정
- 현재 리스트에 값을 출력하는 부분까지 구현완료
- Gson Converter에서 Date 객체는 yyyy-MM-dd HH:mm:ss의 포맷으로 설정
- DB 모델링 수정

+++ app/src/main/java/com/example/margat/controller/PostController.kt
@@ -57,13 +57,15 @@ class PostController {
         var request = RetrofitAPI().creater.create(PostingRequest::class.java)
         request.writePostWithPhotos(map, imageList).enqueue(object: MyCallback<ResponseBody>() {
             override fun onResponse(call: Call<ResponseBody>, response: Response<ResponseBody>) {
-                deleteAllFiles()
+                if (response.code() == 200) {
+                    deleteAllFiles()
 
-                mPostingFragment.clearInputs()
-                mImagePagerAdapter.clearImages()
-                mImagePagerAdapter.notifyDataSetChanged()
+                    mPostingFragment.clearInputs()
+                    mImagePagerAdapter.clearImages()
+                    mImagePagerAdapter.notifyDataSetChanged()
 
-                Toast.makeText(mActivity, "성공적으로 포스팅이 완료되었습니다!", Toast.LENGTH_SHORT).show()
+                    Toast.makeText(mActivity, "성공적으로 포스팅이 완료되었습니다!", Toast.LENGTH_SHORT).show()
+                }
             }
         })
     }
